requirements.txt
Flask
faster-whisper
librosa
numpy
imageio-ffmpeg
textblob
PyPDF2
python-docx
gunicorn
psycopg2-binary
